backend/requirements.txt
fastapi
uvicorn
replicate
python-dotenv
ffmpeg-python
requests
google-generativeai